requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
deprecation==2.1.0
executing==2.1.0
frozendict==2.4.4
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.11.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
postgrest==0.18.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.27.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==308
pyzmq==26.2.0
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
selenium==4.23.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.34
stack-data==0.6.3
storage3==0.9.0
streamlit==1.38.0
streamlit-modal==0.1.2
streamlit-option-menu==0.3.13
supabase==2.10.0
supafunc==0.7.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.41
